backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
pymysql==1.1.1
sqlalchemy==2.0.35

# Zhipu AI
zhipuai>=2.1.0

# DeepSeek AI (使用OpenAI兼容的API)
openai>=1.0.0

# Data processing
pandas==2.2.3
numpy==2.1.3

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# CORS
aiohttp==3.10.10

# Redis (Phase 1 - Cache)
redis>=5.0.0
